requirements.txt
langchain_community
langchain_core
python-dotenv
streamlit
langchain
openai
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdfpip
chromadb
faiss-cpu  
arxiv
wikipedia
langchainhub
groq
cassio
beautifulsoup4
langchain_groq
langgraph